Rename test and add better cookie handling when cookie is not present
